-Moved Image stream fetching and Caching logic to Sample.cs. This way it can also share the HttpClient with the rest of the Sample.
-Fixed casing of return variable for GetDocumentationAsync.

-Use dispose instead of using for SaveImageToCache.
