+++ Microsoft.Toolkit.Uwp.SampleApp/Models/Sample.cs
@@ -25,6 +25,8 @@ using Microsoft.Toolkit.Uwp.Helpers;
 using Microsoft.Toolkit.Uwp.SampleApp.Models;
 using Microsoft.Toolkit.Uwp.UI.Controls;
 using Windows.Foundation.Metadata;
+using Windows.Storage;
+using Windows.Storage.Streams;
 using Windows.UI.Xaml;
 
 namespace Microsoft.Toolkit.Uwp.SampleApp
@@ -148,7 +150,7 @@ namespace Microsoft.Toolkit.Uwp.SampleApp
         }
 
 #pragma warning disable SA1009 // Doesn't like ValueTuples.
-        public async Task<(string contents, string Path)> GetDocumentationAsync()
+        public async Task<(string contents, string path)> GetDocumentationAsync()
 #pragma warning restore SA1009 // Doesn't like ValueTuples.
         {
             if (!string.IsNullOrWhiteSpace(_cachedDocumentation))
@@ -230,6 +232,79 @@ namespace Microsoft.Toolkit.Uwp.SampleApp
             }
 
             return (_cachedDocumentation, _cachedPath);
+        }
+
+        /// <summary>
+        /// Gets the image data from a Uri, with Caching.
+        /// </summary>
+        /// <param name="uri">Image Uri</param>
+        /// <returns>Image Stream</returns>
+        public async Task<IRandomAccessStream> GetImageStream(Uri uri)
+        {
+            async Task<Stream> CopyStream(HttpContent source)
+            {
+                var stream = new MemoryStream();
+                await source.CopyToAsync(stream);
+                stream.Seek(0, SeekOrigin.Begin);
+                return stream;
+            }
+
+            IRandomAccessStream imageStream = null;
+            var localpath = $"{uri.Host}/{uri.LocalPath}";
+
+            // Cache only in Release
+#if !DEBUG
+                try
+                {
+                    imageStream = await StreamHelper.GetLocalCacheFileStreamAsync(localpath, Windows.Storage.FileAccessMode.Read);
+                }
+                catch
+                {
+                }
+#endif
+
+            if (imageStream == null)
+            {
+                try
+                {
+                    using (var response = await client.GetAsync(uri))
+                    {
+                        if (response.IsSuccessStatusCode)
+                        {
+                            var imageCopy = await CopyStream(response.Content);
+                            imageStream = imageCopy.AsRandomAccessStream();
+
+                            // Cache only in Release
+#if !DEBUG
+                                // Takes a second copy of the image stream, so that is can save the image data to cache.
+                                SaveImageToCache(localpath, await CopyStream(response.Content));
+#endif
+                        }
+                    }
+                }
+                catch
+                {
+                }
+            }
+
+            return imageStream;
+        }
+
+        private async void SaveImageToCache(string localpath, Stream imageStream)
+        {
+            var folder = ApplicationData.Current.LocalCacheFolder;
+            localpath = Path.Combine(folder.Path, localpath);
+
+            // Resort to creating using traditional methods to avoid iteration for folder creation.
+            Directory.CreateDirectory(Path.GetDirectoryName(localpath));
+
+            using (var filestream = File.Create(localpath))
+            {
+                await imageStream.CopyToAsync(filestream);
+            }
+
+            // Close copied stream when finished.
+            imageStream.Dispose();
         }
 
         private string ProcessDocs(string docs)
